+++ packages/codeblock/src/config.ts
@@ -1,5 +1,6 @@
 import type { InjectionKey } from 'vue';
 import type { CodeBlockLanguageRuntimeOptions } from './types';
 
-export const CODE_BLOCK_LANGUAGE_OPTIONS_KEY: InjectionKey<CodeBlockLanguageRuntimeOptions> =
-  Symbol('vf-codeblock-language-options');
+export const CODE_BLOCK_LANGUAGE_OPTIONS_KEY: InjectionKey<CodeBlockLanguageRuntimeOptions> = Symbol(
+  'vf-codeblock-language-options',
+);