+++ packages/core/src/components/field/VfField.vue
@@ -88,14 +88,16 @@ provide(vfFieldContextKey, {
 <template>
   <div :class="rootClasses">
     <label v-if="hasLabel && !isFloatingLabel" :for="controlId" class="vf-field__label">
-      <slot name="label">{{ props.label }}</slot><span v-if="props.required" class="vf-field__required-mark" aria-hidden="true">*</span>
+      <slot name="label">{{ props.label }}</slot
+      ><span v-if="props.required" class="vf-field__required-mark" aria-hidden="true">*</span>
     </label>
 
     <div :class="controlClasses">
       <slot :control-id="controlId" :described-by="describedBy" :invalid="isInvalid" :required="props.required" />
 
       <label v-if="hasLabel && isFloatingLabel" :for="controlId" class="vf-field__label vf-field__label--floating">
-        <slot name="label">{{ props.label }}</slot><span v-if="props.required" class="vf-field__required-mark" aria-hidden="true">*</span>
+        <slot name="label">{{ props.label }}</slot
+        ><span v-if="props.required" class="vf-field__required-mark" aria-hidden="true">*</span>
       </label>
     </div>